minor fixes but main bug still present
git-svn-id: https://pure-data.svn.sourceforge.net/svnroot/pure-data/trunk@4166 df7eb9d6-776f-4174-a42d-0ffd8d5d5b5d

+++ externals/frankenstein/rhythms_memory.c
@@ -96,7 +96,7 @@ void end_measure(t_rhythms_memory *x)
 	counter = 0;
 	while(currEvent)
 	{
-		fduration = currEvent->when / x->measure_length;
+		fduration = (float) (((float) currEvent->when) / ((float) x->measure_length));
 		if (x->seq_initialized)
 		{
 			concatenateBeat(x->curr_seq, currEvent->voice, fduration, 1);
